rework python-can transport interface
- move all abstract to AbstractCanTransportInterface
- rework unit and performance tests
